chore(general): Move StatusSyntaxHighlighter to StatusQ, remove mock from Storybook
