+++ storybook/CMakeLists.txt
@@ -35,12 +35,14 @@ set(STATUSQ_HEADERS
     ${STATUSQ_DIR}/include/StatusQ/statuswindow.h
     ${STATUSQ_DIR}/include/StatusQ/typesregistration.h
     ${STATUSQ_DIR}/include/StatusQ/QClipboardProxy.h
+    ${STATUSQ_DIR}/include/StatusQ/statussyntaxhighlighter.h
 )
 
 set(STATUSQ_SOURCES
     ${STATUSQ_DIR}/src/statuswindow.cpp
     ${STATUSQ_DIR}/src/typesregistration.cpp
     ${STATUSQ_DIR}/src/QClipboardProxy.cpp
+    ${STATUSQ_DIR}/src/statussyntaxhighlighter.cpp
 )
 
 if(APPLE)
