--- storybook/mocks/DotherSide/StatusSyntaxHighlighter.qml
@@ -1,7 +1,0 @@
-import QtQuick 2.14
-
-QtObject {
-    property var quickTextDocument
-    property color codeBackgroundColor
-    property color codeForegroundColor
-}
