--- storybook/mocks/DotherSide/qmldir
@@ -1,1 +1,0 @@
-StatusSyntaxHighlighter 0.1 StatusSyntaxHighlighter.qml
