+++ ui/StatusQ/include/StatusQ/statussyntaxhighlighter.h
