+++ ui/StatusQ/src/typesregistration.cpp
@@ -1,7 +1,8 @@
 #include "StatusQ/typesregistration.h"
 
-#include "StatusQ/statuswindow.h"
 #include "StatusQ/QClipboardProxy.h"
+#include "StatusQ/statussyntaxhighlighter.h"
+#include "StatusQ/statuswindow.h"
 
 #include <QQmlEngine>
 
@@ -10,4 +11,5 @@ void registerStatusQTypes()
 	qmlRegisterType<StatusWindow>("StatusQ", 0 , 1, "StatusWindow");
     qmlRegisterSingletonType<QClipboardProxy>("StatusQ", 0 , 1, "QClipboardProxy",
                                               &QClipboardProxy::qmlInstance);
+    qmlRegisterType<StatusSyntaxHighlighter>("StatusQ", 0 , 1, "StatusSyntaxHighlighter");
 }
